Separate SCL and SDA Clocks
Needed to shift the scl clock 45 deg, in order to ensure we prevent
accidental start or stop conditions (e.g. when sda changes, we must have
scl=0).

We needed to add a cycle where scl was low before The stop condition in
order to guard against accidental start condition there.

9 bits are bieng created, everything is still at approximately 100KHz in
this scheme, simplifying the implementation.